refactor(listen): native barge/speaker/ASR gates bound at boot (no phase labels)
