+++ sensory/voice_gates.py
@@ -1,143 +1,17 @@
-"""Shared barge / speaker gate flags + install hooks for AikoListen."""
-from __future__ import annotations
-
-import os
-import time
-from typing import Any
-
-_LOG = None
-_HOOKS_INSTALLED = False
+"""Backward-compatible re-exports from sensory.listen_native."""
+from sensory.listen_native import (  # noqa: F401
+    barge_in_always_on,
+    barge_in_enabled,
+    bind_native_gates,
+    speaker_verify_gate,
+)
 
 
-def _log():
-    global _LOG
-    if _LOG is None:
-        import logging
-        _LOG = logging.getLogger(__name__)
-    return _LOG
+def install_listen_hooks(listen=None) -> None:
+    """Compat: bind native gates (idempotent)."""
+    bind_native_gates()
 
 
-def _env_bool(name: str, default: str = "0") -> bool:
-    return os.getenv(name, default).strip().lower() in {"1", "true", "yes", "on"}
-
-
-def barge_in_enabled() -> bool:
-    """Master barge-in switch (browser + Jetson)."""
-    return _env_bool("BARGE_IN_ENABLED", "0")
-
-
-def barge_in_always_on() -> bool:
-    """Local Silero monitor outside TTS wait — only if barge_in_enabled()."""
-    return barge_in_enabled() and _env_bool("BARGE_IN_ALWAYS_ON", "0")
-
-
-def speaker_verify_enabled() -> bool:
-    return _env_bool("SPEAKER_VERIFY_ENABLED", "0")
-
-
-def speaker_verify_gate() -> bool:
-    """When True with verify active, failed match drops the utterance."""
-    return _env_bool("SPEAKER_VERIFY_GATE", "0")
-
-
-def install_listen_hooks(listen: Any = None) -> None:
-    """Patch AikoListen barge + speaker-gate + ASR text correct (idempotent)."""
-    global _HOOKS_INSTALLED
-    if _HOOKS_INSTALLED:
-        return
-    try:
-        from sensory import listen as mod
-    except Exception as e:
-        _log().warning("voice gate hooks deferred: %s", e)
-        return
-
-    cls = mod.AikoListen
-
-    _orig_trigger = cls.trigger_barge_in
-
-    def trigger_barge_in(self) -> None:
-        if not barge_in_enabled():
-            return
-        _orig_trigger(self)
-
-    cls.trigger_barge_in = trigger_barge_in
-
-    _orig_listen = cls.listen
-
-    def listen_wrapped(
-        self,
-        status_callback=None,
-        wait_fn=None,
-        speak=None,
-        chunk_source=None,
-        vad_presegmented: bool = False,
-    ):
-        waited_tts = False
-        if speak is not None and speak.is_playing() and not barge_in_enabled():
-            if status_callback:
-                try:
-                    status_callback("__WAITING__")
-                except Exception:
-                    pass
-            while speak.is_playing():
-                time.sleep(0.05)
-            speak = None
-            waited_tts = True
-
-        text, info = _orig_listen(
-            self,
-            status_callback=status_callback,
-            wait_fn=None if waited_tts else wait_fn,
-            speak=speak,
-            chunk_source=chunk_source,
-            vad_presegmented=vad_presegmented,
-        )
-        if (
-            speaker_verify_gate()
-            and self.speaker_verify_active()
-            and info is not None
-            and info.get("verified") is False
-        ):
-            _log().info(
-                "[gate] speaker verify failed (score=%s) — dropping utterance",
-                info.get("speaker_score"),
-            )
-            return "", info
-        return text, info
-
-    cls.listen = listen_wrapped
-
-    _orig_loop = cls._barge_in_loop
-
-    def _barge_in_loop(self) -> None:
-        if not barge_in_enabled():
-            while getattr(self, "_barge_in_active", False):
-                time.sleep(0.5)
-            return
-        _orig_loop(self)
-
-    cls._barge_in_loop = _barge_in_loop
-
-    try:
-        from sensory.asr_correct import correct_asr_text
-        if not getattr(cls, "_asr_correct_installed", False):
-            orig = cls._transcribe
-
-            def _transcribe(self, audio):
-                text = orig(self, audio)
-                try:
-                    return correct_asr_text(text)
-                except Exception:
-                    return text
-
-            cls._transcribe = _transcribe
-            cls._asr_correct_installed = True
-    except Exception:
-        _log().exception("ASR text-correct hooks failed")
-
-    _HOOKS_INSTALLED = True
-    _log().info("voice gate hooks installed (barge / speaker gate / ASR text correct)")
-
-
-def install_listen_s0_hooks(listen: Any = None) -> None:
-    install_listen_hooks(listen)
+def install_listen_s0_hooks(listen=None) -> None:
+    """Compat alias."""
+    bind_native_gates()
